Review item: Remove models.go and move the model structs into the modules where they are used.
