+++ tharsis/data_source_workspace_outputs.go
@@ -22,6 +22,15 @@ const (
 	tharsisGroupPathEnvVar = "THARSIS_GROUP_PATH"
 )
 
+// WorkspacesOutputsDataSourceData represents the outputs for a workspace in Tharsis.
+type WorkspacesOutputsDataSourceData struct {
+	Outputs        map[string]string `tfsdk:"outputs"`
+	Path           types.String      `tfsdk:"path"`
+	FullPath       types.String      `tfsdk:"full_path"`
+	WorkspaceID    types.String      `tfsdk:"workspace_id"`
+	StateVersionID types.String      `tfsdk:"state_version_id"`
+}
+
 // Ensure provider defined types fully satisfy framework interfaces
 var (
 	_ datasource.DataSource = workspaceOutputsDataSource{}
